+++ lib/src/nonlinear/nonlinear.dart
@@ -18,8 +18,8 @@ import 'package:equations/equations.dart';
 ///   - [Brent]
 ///   - [RegulaFalsi]
 ///
-/// Each subclass of [NonLinear] has to define the [solve] method which is required
-/// in order to build the scalar succession with a certain logic.
+/// Each subclass of [NonLinear] has to define the [solve] method which is
+/// required in order to build the scalar succession with a certain logic.
 abstract class NonLinear {
   /// The function f(x) for which the algorithm has to find a solution.
   final String function;
@@ -114,11 +114,6 @@ abstract class NonLinear {
 
     return (upper - lower) / (2 * h);
   }
-
-  /// Calculates the numerical value of the integral of this function using a
-  /// [NumericalIntegration] algorithm.
-  IntegralResults integrateOn(NumericalIntegration numericalIntegration) =>
-      numericalIntegration.integrate(function);
 
   /// Returns a [NonlinearResults] object which contains the data calculated by
   /// the root-finding algorithm.